+++ MathGamesAPI/app/controllers/notification.controller.js
@@ -1,5 +1,6 @@
 const db = require("../models");
 const Notification = db.notifications;
+const User = db.user;
 const Op = db.Sequelize.Op;
 
 // Create and Save a new Notification
@@ -12,29 +13,46 @@ exports.create = (req, res) => {
     return;
   }
 
-  let date = new Date();
-  date.setTime( date.getTime() - new Date().getTimezoneOffset()*60*1000 );
-  date = date.toISOString().slice(0, 19).replace('T', ' ');
-  
+  if (req.body.sender !== req.userId) {
+    res.status(401).send({
+      message: "Unauthorized!"
+    });
+    return;
+  }
+
   // Create a Notification
   const notification = {
     sender: req.body.sender,
     receiver: req.body.receiver,
-    notification_type: req.body.notification_type,
-    notification_date: date
+    notification_type: req.body.notification_type
   };
 
-  // Save Notification in the database
-  Notification.create(notification)
-    .then(data => {
-      res.send(data);
-    })
-    .catch(err => {
-      res.status(500).send({
-        message:
-          err.message || "Some error occurred while creating the Notification."
+  Notification.findOne({where: {
+    sender: notification.sender,
+    receiver: notification.receiver,
+    notification_type: notification.notification_type
+  }}).then(response => {
+    if (response !== null) {
+      res.status(405).send({
+        message: "Friend Request already made."
       });
+    } else {
+      Notification.create(notification)
+      .then(data => {
+        res.send(data);
+      })
+      .catch(err => {
+        res.status(500).send({
+          message: err.message || "Some error occurred while creating the Notification."
+        });
+      });
+    }
+  }).catch(err => {
+    res.status(500).send({
+      message: err.message || "Some error occurred while searching for Notification Duplication."
     });
+  })
+
 };
 
 // Retrieve all Notifications from the database.
@@ -45,8 +63,7 @@ exports.findAll = (req, res) => {
     })
     .catch(err => {
       res.status(500).send({
-        message:
-          err.message || "Some error occurred while retrieving Notifications."
+        message: err.message || "Some error occurred while retrieving Notifications."
       });
     });
 };
@@ -55,7 +72,16 @@ exports.findAll = (req, res) => {
 exports.findByUserId = (req, res) => {
   const id = req.params.id;
 
-  Notification.findAll({where: {receiver: id}})
+  if (parseInt(req.userId) !== parseInt(id)) {
+    res.status(401).send({
+      message: "Unauthorized!"
+    });
+    return;
+  }
+
+  Notification.findAll({attributes: ['id', 'receiver', 'notification_type', 'createdAt' ], 
+                        where: {receiver: id}, include: [{model: User, as: 'sender_user', attributes: [['id', 'sender_id'], ['username', 'sender']], 
+                        order: [["createdAt", 'DESC']] }]})
     .then(data => {
       res.send(data);
     })
@@ -70,25 +96,48 @@ exports.findByUserId = (req, res) => {
 exports.delete = (req, res) => {
   const id = req.params.id;
 
-  Notification.destroy({
-    where: { id: id }
-  })
-    .then(num => {
-      if (num == 1) {
-        res.send({
-          message: "Notification was deleted successfully!"
-        });
-      } else {
-        res.send({
-          message: `Cannot delete Notification with id=${id}. Maybe Notification was not found!`
-        });
-      }
-    })
-    .catch(err => {
-      res.status(500).send({
-        message: "Could not delete Notification with id=" + id
+
+  Notification.findOne({where: {id: id}}).then(notif => {
+    if (notif === null) {
+      res.status(404).send({
+        message: "Notification not found"
       });
-    });
+      return;
+    }
+    
+    if ( (parseInt(notif.receiver) !== parseInt(req.userId)) && (parseInt(notif.sender) !== parseInt(req.userId)) ) {
+      res.status(401).send({
+        message: "Unauthorized!"
+      });
+      return;
+
+    } else {
+      Notification.destroy({
+        where: { id: id }
+      })
+        .then(num => {
+          if (num == 1) {
+            res.send({
+              message: "Notification was deleted successfully!"
+            });
+          } else {
+            res.status(500).send({
+              message: `Cannot delete Notification with id=${id}. Maybe Notification was not found!`
+            });
+          }
+        })
+        .catch(err => {
+          res.status(500).send({
+            message: "Could not delete Notification with id=" + id
+          });
+        });
+
+    }
+  }).catch(err => {
+    res.status(500).send({
+        message:  err.message || "Some error occurred while removing Notification."
+    })
+});
 };
 
 // Delete all Notifications from the database.
